+++ src/components/templates/admin/pointConfiguration/services/service-add/index.tsx
@@ -1,9 +1,6 @@
 import { Box, Flex, Spacer, Text } from "@chakra-ui/react";
 import FormControl from "@src/components/atoms/FormControl";
-import { getAllMemberTier } from "@src/service/master-data/member-tier";
 import { colors } from "@src/theme/colors";
-import { useEffect } from "react";
-import { useQuery } from "react-query";
 import styled from "styled-components";
 const Wrapper = styled.div`
   display: flex;
@@ -17,28 +14,33 @@ const Wrapper = styled.div`
 export const CreateServiceForm = ({
   register,
   errors,
-  formDataArray,
   setFormDataArray,
   setValue,
   watch,
 }: any) => {
-  const { data } = useQuery("member_tier", getAllMemberTier, {
-    select: ({ data }) => data.datalist,
-  });
-  const handleInputChange = (e: any, index: number, id: number) => {
-    const { value } = e.target;
-    const newFormDataArray = [...formDataArray];
-    newFormDataArray[index] = {
-      ...newFormDataArray[index],
-      id: id,
-      rewardPercentage: value,
-    };
-
-    setFormDataArray(newFormDataArray.filter((item) => item !== undefined));
+  const updatePercentValue = (
+    array: any,
+    idToUpdate: any,
+    newPercentValue: any
+  ) => {
+    return array.map((item: any) => {
+      if (item.id === idToUpdate) {
+        return { ...item, rewardPercentage: newPercentValue };
+      }
+      return item;
+    });
   };
-  useEffect(() => {
-    setValue("membershipServiceResponseDtos", data);
-  }, [data]);
+
+  const handleInputChange = (e: any, item: any) => {
+    const { value } = e.target;
+    const dat = updatePercentValue(
+      watch("membershipServiceResponseDtos"),
+      item.id,
+      value
+    );
+    setFormDataArray(dat);
+    setValue("membershipServiceResponseDtos", dat);
+  };
   return (
     <>
       <Box mx={{ base: "none", md: "auto" }}>
@@ -75,12 +77,12 @@ export const CreateServiceForm = ({
                       type="text"
                       name={`rewardPercentage${index}`}
                       register={register}
-                      value={itm.rewardPercentage}
+                      defaultValue={itm.rewardPercentage}
                       placeholder={""}
                       label={itm.membershipName}
                       error={errors?.membershipName?.message || ""}
                       label_color={colors.black_1}
-                      onChange={(e: any) => handleInputChange(e, index, itm.id)}
+                      onChange={(e: any) => handleInputChange(e, itm)}
                     />
                   );
                 }