+++ src/interface/master-data/property/index.tsx
@@ -1,13 +1,21 @@
 export interface IProperty {
   id?: number | string;
-  name: string;
-  code: string;
-  phoneNumber: string;
-  contactPerson: string;
-  contactPersonPhoneNo: string;
+  name?: string;
+  code?: string;
+  phoneNumber?: string;
+  contactPerson?: string;
+  contactPersonPhoneNo?: string;
 }
 
 export interface IMemberTier {
-  data: string;
-  image: string;
+  data?: string;
+  image?: string;
+}
+
+export interface IMemberTierDetail {
+  id?: number | string;
+  membershipName?: string;
+  imageUrl?: string;
+  image?: string | undefined | Blob;
+  requiredPoints?: number | string;
 }