+++ src/pages/PointConfig/Service/index.tsx
@@ -1,7 +1,7 @@
 import { BreadCrumb } from "@src/components/atoms/Breadcrumb";
 import Content from "@src/components/molecules/content";
 import ServiceList from "@src/components/templates/admin/pointConfiguration/services/service-list";
-import { getAllService } from "@src/service/point-config";
+import { getAllService } from "@src/service/point-config/service";
 import { useQuery } from "react-query";
 
 const ServicePage = () => {
@@ -12,7 +12,7 @@ const ServicePage = () => {
     <>
       <BreadCrumb name="Point Configuration" />
       <Content>
-        <ServiceList data={data} isLoading={isLoading} />
+        <ServiceList tableData={data} tableDataFetching={isLoading} />
       </Content>
     </>
   );